revert probe pkg introduction, to avoid unnecessary public types and methods
